requirements.txt
onnxruntime
numpy
pillow
fastapi
uvicorn
torch
torchvision